terraform/src/eliza_auth_login/requirements.txt
pydantic
pydantic-core
base58
solders
boto3
pyjwt